Deleted the Machine class and created a Box class
Deleted the Machine class that I was using and replaced it with a class
called Box that focuses on the boxes, rather than the machines.
Added a new class called Mech that I will work on. That's the previous
Machine class that has to be started from scratch.
Made the box objects automatically get the position where they should
spawn, rather than having a hardcoded point to start from.