:memo: Big changes. Read desc
:crying_cat_face: So I figured out a really big bug that was giving me
problems. In order to solve that I had to change many things in the
code.
I was using x and y in the classes thinking that they were the indexes
for rows and cols. It turns out that I made x the index for cols and y
the index for rows ( which makes sense if you think about it since y is
the vertical axis )
So by mistake I had those switched around and always gave me a big pain
when I was changing values in the array.
I removed the previously created c_array_boxes because it was not needed
anymore.
With that I also removed the initialisation of the array.
:collision:  Probably the most important things is that I added the
collision with the boxes. Now you won't be able to go left or right if
you have boxes on those positions.
Also added two if statements that update the character's position (x and
y values)
Added a test tool that teleports the player to the location I click on
Many more small changes